recycling_num=1 只选取 '1wcq','5u7m','2nno','5j87','5uwf','5kxi','4kju','3eta','2b54','1sdt' 作为训练样本
加入当前构象到真实构象的旋转平移扭转量，直接有监督学习
RMSD准确公式

+++ tankbind/utils.py
@@ -563,7 +563,7 @@ class OptimizeConformer:
                                         maxiter=maxiter, popsize=popsize,
                                         mutation=mutation, recombination=recombination, disp=False, seed=seed)
         opt_rmsd,opt_R, opt_tr=self.apply_torsion(result["x"])
-        opt_torsion=torch.from_numpy(result["x"]).to(self.candidate_pos.device)
+        opt_torsion=torch.from_numpy(result["x"]).to(self.candidate_pos.device).float()
         opt_rotate=matrix_to_axis_angle(opt_R).float()
         opt_tr=opt_tr.T[0]
         return opt_tr,opt_rotate,opt_torsion,opt_rmsd
